requirements.txt
torch>=2.1.0
torchvision>=0.16.0
Pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.66.0
matplotlib>=3.7.0
